bootsrap and base.html added

+++ django/todo_list/base/views.py
@@ -12,9 +12,13 @@ def index(request):
 
 def taskList(request):
 
-    form = TaskForm()
+    if request.method == 'GET':
 
-    tasks = Task.objects.all()
+        form = TaskForm()
+
+        tasks = Task.objects.all()
+
+        context = {'tasks': tasks, 'TaskForm': form}
 
     if request.method == 'POST':
         form = TaskForm(request.POST)
@@ -28,21 +32,22 @@ def taskList(request):
 
             return redirect('/tasklist')
 
-    context = {'tasks': tasks, 'TaskForm': form}
-
     return render(request, 'base/tasklist.html', context)
 
 
 def updateTask(request, pk):
 
-    task = Task.objects.get(id=pk)
+    if request.method == 'GET':
 
-    form = TaskForm(instance=task)
+        task = Task.objects.get(id=pk)
 
-    print("Update form loaded.")
+        form = TaskForm(instance=task)
+
+        print("Update form loaded.")
+        context = {'TaskForm': form}
 
     if request.method == 'POST':
-
+        task = Task.objects.get(id=pk)
         form = TaskForm(request.POST, instance=task)
 
         if form.is_valid():
@@ -51,17 +56,19 @@ def updateTask(request, pk):
             print("Update executed.")
             return redirect('/tasklist')
 
-    context = {'TaskForm': form}
     return render(request, 'base/updatetask.html', context)
 
 
 def deleteTask(request, pk):
-    print("Delete form opened.")
-    task = Task.objects.get(id=pk)
+    if request.method == 'GET':
+        print("Delete form opened.")
+        task = Task.objects.get(id=pk)
+        context = {'task': task}
+        return render(request, 'base/deletetask.html', context)
+
     if request.method == 'POST':
+        task = Task.objects.get(id=pk)
+        print("POST PK:"+pk)
         print("Delete executed.")
         task.delete()
         return redirect('tasklist')
-
-    context = {'task': task}
-    return render(request, 'base/deletetask.html', context)
